Add empty CI job in preparation for unit tests (#1)
Summary:
This PR just adds a basic CI job that does nothing apart from installing pytorch. We need to merge it first for the CI job to actually trigger on subsequent PRs, at which point I'll be able to start adding the actual package build and running the tests.

Pull Request resolved: https://github.com/pytorch-labs/torchcodec/pull/1

Reviewed By: scotts

Differential Revision: D57967539

Pulled By: NicolasHug

fbshipit-source-id: 15417ebae19fac1b4bfd8da5ffeed83414729387