[torchcodec] Get C++ VideoDecoderTest to build and run in OSS
Summary:
1. Split up the CMakeLists.txt into 3 files -- a main file, a library file and a test file. This is because we don't want to rebuild the library only for testing and instead only refer to it.
2. Remove .bmp reference files and a dependency on opencv. This is because libtorch is typically compiled with without cxx11 ABI while the precompiled binaries of opencv are compiled with cxx11 ABI. These libraries are incompatible with each other.
This also cuts the compilation time as there are fewer dependencies. We can easily convert .pt files to .bmp files using a few lines of python code.
4. Guard some internal things using #ifdef macros.

Test plan shows that we are able to build and test the code. 2 tests are broken likely because of my ffmpeg version not matching fb's ffmpeg version -- we can investigate that later.

TODO: Get VideoDecoderOpsTest to work as well and make the 2 remaining testcases pass.

Reviewed By: NicolasHug

Differential Revision: D57912312

fbshipit-source-id: 64f6182a6f84f4af0bf57908f468128581c33900