Build and run C++ tests on GitHub CI (#12)
Summary:
This PR lets the GitHub CI build and run the C++ tests.

https://github.com/pytorch-labs/torchcodec/pull/13 and https://github.com/pytorch-labs/torchcodec/pull/14 show that CI is expectedly red when either the Python or the C++ test fail.

We're building `libtorchcodec.so` twice: once during the installation of the Python package, and a second time added here when building the tests. This is OK for now as it's fast, but not ideal and we should find a way to only build once. I added a TODO for that.

Pull Request resolved: https://github.com/pytorch-labs/torchcodec/pull/12

Reviewed By: scotts

Differential Revision: D58186092

Pulled By: NicolasHug

fbshipit-source-id: eada1692445d9520c22d0f99cb7944acf44fae41