Test Python{3.8, 3.12} X ffmpeg{4, 5, 6, 7} (#16)
Summary:
Now that we support ffmpeg 4 to 6 this PR let the GitHub CI run the tests on all of these ffmpeg versions. We are restricting the Python versions to 3.8 and 3.12 (at least for now) to avoid too many jobs. Closer to the release we might want to increase coverage.

Pull Request resolved: https://github.com/pytorch-labs/torchcodec/pull/16

Reviewed By: ahmadsharif1

Differential Revision: D58192553

Pulled By: NicolasHug

fbshipit-source-id: 06416baddbf0f849908c19ceb38151a712a90e40