More robust wheel build (#32)
Summary:
In `main`, we're building the wheels with the `python -m build . --wheel` command. The recommended way to build a wheel is in fact to invoke `python -m build .` without the `--wheel` part. This is more robust, because this will first build a source dist, and then build the wheel from that source dist.

More details form the `build` help:

>  By default, a source distribution (sdist) is built from {srcdir} and a binary distribution (wheel) is built from the sdist. This is recommended as it will ensure the sdist can be used to build wheels.
    Pass -s/--sdist and/or -w/--wheel to build a specific distribution. If you do this, the default behavior will be disabled [...]

Prior to this PR, building a source dist would break. We need to specify which files goes into the source dist, and this is done via the newly added `MANIFEST.in` file.

Pull Request resolved: https://github.com/pytorch-labs/torchcodec/pull/32

Reviewed By: scotts

Differential Revision: D58646827

Pulled By: NicolasHug

fbshipit-source-id: c3af44a1e83d96b4bc34f55d8be42aa39f50d95f