[torchcodec] unify stream_index placement in core API (#49)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/torchcodec/pull/49

We initially agree on placing `stream_index` last in the core APIs, but that was based on a mistaken understanding that it was optional. Since it is not optional, it makes more sense for it to appear early, always after the decoder. It still remains keyword-only, as it is an integer parameter among many others.

Differential Revision: D58952296