Use zip instead of enumerate (#55)
Summary:
Minor code quality improvement.

Maybe the original use of `enunmerate()` was intentional? Happy to reconsider if that's the case.

Otherwise, `zip()` is more commonly use for iterating over pairs. Another benefit of `zip` here is robustness: We are 100% sure the indexing works the same on our video decoder as well as on the reference lists (e.g. `ref_frames0_9[4:8]`), instead of doing the index arithmetic ourselves.

Pull Request resolved: https://github.com/pytorch-labs/torchcodec/pull/55

Reviewed By: scotts, ahmadsharif1

Differential Revision: D59107586

Pulled By: NicolasHug

fbshipit-source-id: ddfbe6894062a871be08a42984063cae8393a3ed