[torchcodec] Add a core function to get the frame along with metadata like pts and duration (#61)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/torchcodec/pull/61

Before this diff there wasn't a way for python users to get the frame metadata like pts and duration.

This diff adds a function to get the frame, pts and duration for a frame at a given index.

TODO: Add similar functions for get_next_frame and get_frame_at_pts.

Differential Revision: D59328760