[torchcodec] Revert register_fake back to imply_abstract (#69)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/torchcodec/pull/69

`register_fake` is only available after Pytorch 2.4, which will be available in late July. We don't want to break user experience if they on old pytorch versions. It is also possible when we release torchcodec, pytorch 2.4 not yet released.

Note: it's expected that reverting back will trigger a lot of logging.

Reviewed By: NicolasHug

Differential Revision: D59018191