[torchcodec] Update functions that return a single frame to return frame and metadata (#74)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/torchcodec/pull/74

We now return the frame tensor, frame pts and frame duration for these functions.

Users who use the sampler or SimpleVideoDecoder are not affected.

Note that the pts and duration are returned as Tensors. This is due to a limitation of torch.compile() where it will not accept tuples of non-Tensors:

https://fb.workplace.com/groups/1075192433118967/posts/1461960124442194/?comment_id=1461971881107685

I also removed the `get_frame_with_info_at_index` because now that's redundant with `get_frame_at_index`.

In the proceses I also included VideoDecoderOps.h in VideoDecoderOps.cpp and fixed some inconsistencies between the two that were causing compilation failures.

TODO: I will handle the batch API functions in a subsequent diff.

Reviewed By: scotts, NicolasHug

Differential Revision: D59542194