Build wheel on gcc 9 (like pytorch) and add sanity check for wheel (#76)
Summary:
This PR ensures that we build our wheel on gcc 9 (like torch) and adds a sanity check for wheel compatibility. This should prevent a common problem that ahmadsharif1 and I have often observed when building locally. Please see comments in the code for a full explanation of the issue we're addressing.

Pull Request resolved: https://github.com/pytorch-labs/torchcodec/pull/76

Reviewed By: scotts, ahmadsharif1

Differential Revision: D59671770

Pulled By: NicolasHug

fbshipit-source-id: 2beb22af9fdc3df026e39c43d0693952fdd537fa