[torchcodec] refactor stream index and scanned checks into validate functions (#83)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/torchcodec/pull/83

We've been doing in-place checks for `streamIndex` validation and to validate that we've done a full scan. This leads to copy-paste code and may lead to subtle differences in the validations (such as not using `TORCH_CHECK` and inconsistent messages). This diff consolidates those checks into functions. These validations should only be necessary on the public-facing methods of `VideoDecoder` that have these assumptions.

Reviewed By: NicolasHug, ahmadsharif1

Differential Revision: D59916313