[torchcodec] Add stream_index to the output of core functions (#87)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/torchcodec/pull/87

We need to return stream_index for handling torchaudio style de-multiplexing on the python side where the user could configure 2 streams and asks for the next frame from any available stream and uses that.

I also renamed the tuples to be closer to their names on the C++ side.

Differential Revision: D59930581