Ignore unactionable warning from PyTorch in create_from_bytes (#108)
Summary:
Using `create_from_bytes` generates the following warning:

```
UserWarning: The given buffer is not writable, and PyTorch does not support non-writable tensors. This means you can write to the underlying (supposedly non-writable) buffer using the tensor. You may want to copy the buffer to protect its data or make it writable before converting it to a tensor. This type of warning will be suppressed for the rest of this program. (Triggered internally at ../torch/csrc/utils/tensor_new.cpp:1544.)
  return create_from_tensor(torch.frombuffer(video_bytes, dtype=torch.uint8))
```

This is not actionable for users in this specific context, and also not a problem anyway. This PR silences this warning.

Pull Request resolved: https://github.com/pytorch-labs/torchcodec/pull/108

Reviewed By: scotts

Differential Revision: D60168592

Pulled By: NicolasHug

fbshipit-source-id: ce429d6e002185a26366c2b75d910a2a39255dd5