Revamp notebook example (#110)
Summary:
This PR re-writes our notebook-style example to cover all methods, attributes and related classes of `SimpleVideoDecoder`. And also to show a cute puppy, ~~which will hopefully account for 95% of the success of this project~~.

The example is a bit verbose, by design. I think it's fine, we'll add a much more compact and to-the-point example in the README.

The example currently renders as follows:

![image](https://github.com/user-attachments/assets/8f399f39-f456-4b26-87ed-6b0a4b86ba12)

Pull Request resolved: https://github.com/pytorch/torchcodec/pull/110

Reviewed By: ahmadsharif1

Differential Revision: D60373589

Pulled By: NicolasHug

fbshipit-source-id: 5b8ce234d7a687b98bdb69c1272f12c4983f797f