[torchcodec] Handle negative buffer sizes (#131)
Summary:
Pull Request resolved: https://github.com/pytorch/torchcodec/pull/131

Not sure if we will ever hit this case, but this may be a reason why TorchCodec is crashing on some videos.

Maybe the user passes in a negative value for size and it trickles all the way to the read function. We should not be treating negative sizes as large sizes.

Reviewed By: scotts

Differential Revision: D60461269

fbshipit-source-id: 9172da109d14ddc505177988e85bc173b1fbf403