pytorch/torchcodec - 3.12 fixes (#744)
Summary:

Noticed the abandoned codemod (D77291340) and figured I would fix the test. 
importlib.resources.path has always returned a context manger, but I guess it was pathlike before so nobody bothered using it.   So going back to the introduction of importlib.resources in 3.7 its been a context manager so this is safe.

Differential Revision: D77327183