[TorchCodec] Late-init TorchVision imports
Summary: TorchCodecCoreNonBatch, TorchCodecPublic, and TorchCodecPublicNonBatch define a `decode_and_resize()` function that depends on TorchVision for a resize function, but when using the benchmark utility (`benchmark_decoders.py`) this function never gets called. By moving the TorchCodec imports to a late-init function, we can still use these classes with the benchmark utility when TorchVision isn't available, or TorchVision uses a conflicting version of FFmpeg.

Differential Revision: D80824234