[TorchCodec] Late-init TorchVision imports (#842)
Summary:
Pull Request resolved: https://github.com/pytorch/torchcodec/pull/842

TorchCodecCoreNonBatch, TorchCodecPublic, and TorchCodecPublicNonBatch define a `decode_and_resize()` function that depends on TorchVision for a resize function, but when using the benchmark utility (`benchmark_decoders.py`) this function never gets called. By moving the TorchCodec imports to a late-init function, we can still use these classes with the benchmark utility when TorchVision isn't available, or TorchVision uses a conflicting version of FFmpeg.

Reviewed By: markshw

Differential Revision: D80824234